WIP: Implemented all parsers
There are definitly more unit tests that need to be written/expanded on the
ensure that the parsers are actually working correctly. Additionally the
parsers are just generating arbitrary objects, the next step will be to make
the parsers generate the correct object types to work with the rest of
the bnf library.